refactor(accordion): restructure component architecture and migrate to css modules
Reorganized the Accordion component structure to follow a cleaner architecture pattern
with dedicated subcomponents. Migrated from utility classes to CSS modules for better
encapsulation and maintainability. Removed accordion animations from global CSS and
made them component-specific. Fixed export structure for consistency.
