+++ src/assets/css/theme/animations.css
@@ -1,7 +1,5 @@
 @theme static {
  /* Animations */
- --animate-accordion-down: accordion-down 0.2s ease-out;
- --animate-accordion-up: accordion-up 0.2s ease-out;
  --animate-fade-in: fadein 0.3s ease-out forwards;
  --animate-fade-out: fadeout 0.2s ease-in forwards;
  --animate-language-close: language-close 0.15s ease-in-out;
@@ -10,26 +8,6 @@
  --animate-scale-out: scaleout 0.2s ease-in forwards;
 
  /* Keyframes */
- @keyframes accordion-down {
-  from {
-   height: 0;
-  }
-
-  to {
-   height: var(--radix-accordion-content-height);
-  }
- }
-
- @keyframes accordion-up {
-  from {
-   height: var(--radix-accordion-content-height);
-  }
-
-  to {
-   height: 0;
-  }
- }
-
  @keyframes fadeIn {
   0% {
    opacity: 0;
