+++ src/shared/Accordion/index.ts
@@ -1,0 +1,4 @@
+export { Accordion, type AccordionProperties } from "./ui/Accordion";
+export { AccordionContent } from "./ui/AccordionContent/AccordionContent";
+export { AccordionItem, type AccordionItemProperties } from "./ui/AccordionItem/AccordionItem";
+export { AccordionTrigger } from "./ui/AccordionTrigger/AccordionTrigger";
