+++ src/shared/Accordion/ui/AccordionContent/AccordionContent.module.scss
@@ -1,0 +1,42 @@
+.content {
+ overflow: hidden;
+ transition: all 0.3s ease;
+
+ &[data-state="closed"] {
+  animation: accordion-up 0.2s ease-out;
+ }
+
+ &[data-state="open"] {
+  animation: accordion-down 0.2s ease-out;
+ }
+}
+
+.inner {
+ padding: var(--spacing-5px) var(--spacing-15px) var(--spacing-15px) var(--spacing-15px);
+ font: var(--text-xs);
+
+ @media (width >= 768px) {
+  padding: 0 var(--spacing-20px) var(--spacing-20px) var(--spacing-20px);
+  font: var(--text-sm);
+ }
+}
+
+@keyframes accordion-down {
+ from {
+  height: 0;
+ }
+
+ to {
+  height: var(--radix-accordion-content-height);
+ }
+}
+
+@keyframes accordion-up {
+ from {
+  height: var(--radix-accordion-content-height);
+ }
+
+ to {
+  height: 0;
+ }
+}
